add confirm message for replay, undo add all

+++ src/main/java/page/battle/RecdManagePage.java
@@ -5,6 +5,7 @@ import common.pack.PackData;
 import common.pack.UserProfile;
 import common.util.stage.Replay;
 import main.MainBCU;
+import main.Opts;
 import page.JBTN;
 import page.JTF;
 import page.Page;
@@ -49,8 +50,8 @@ public class RecdManagePage extends AbRecdPage {
 		change(true);
 		Replay r = jlr.getSelectedValue();
 		Vector<Replay> replays = new Vector<>(Replay.getMap().values());
-		for (PackData.UserPack pack : UserProfile.getUserPacks())
-			replays.addAll(pack.getReplays());
+//		for (PackData.UserPack pack : UserProfile.getUserPacks())
+//			replays.addAll(pack.getReplays());
 		jlr.setListData(replays);
 		jlr.setSelectedValue(r, true);
 		setRecd(r);
@@ -77,14 +78,24 @@ public class RecdManagePage extends AbRecdPage {
 			if (isAdj() || jlr.getValueIsAdjusting())
 				return;
 			Replay r = jlr.getSelectedValue();
-			if (r == null)
+			String n = rena.getText();
+			if (r == null || r.rl.id.equals(n))
 				return;
-			r.rename(MainBCU.validate(rena.getText().trim(),'#'), true);
+
+			String name = MainBCU.validate(rena.getText().trim(),'#');
+			if (!Replay.getMap().containsKey(name) || Opts.conf("A replay named " + name + " already exists. Do you wish to overwrite?"))
+				r.rename(name, true);
 			rena.setText(r.rl.id);
 		});
 
 		dele.addActionListener(arg0 -> {
+			if (isAdj() || jlr.getValueIsAdjusting())
+				return;
 			Replay r = jlr.getSelectedValue();
+
+			if (!Opts.conf("Are you sure you want to delete " + r.rl.id + "?"))
+				return;
+
 			File f = CommonStatic.ctx.getWorkspaceFile(r.rl.getPath() + ".replay");
 			if (f.exists()) {
 				f.delete();
